fix: deprecate old use of hyphenators
Problem:
We should not break actual  behaviour of hyphenators.

Solution:
Emit warning when a stringHyphenator gets an HTMLElement to hyphenate,
but still hyphenate the element (but not it's children).

Notes:
With this, the PR can go in a minor release. Major release is to early.
Remove this functionality in next major release.
[skip travis-ci]