Do not hyphenate subtag with different lang in hyphenators (#111)
fix: don't hyphenate subtag with different lang

Problem:
language specific hyphenators try to hyphenate elements with
a different language set. This can cause issues if the engine for
that different language isn't ready, yet.

Solution:
Deprecate old use of hyphenators and emit warning when a
stringHyphenator gets an HTMLElement to hyphenate, but still
hyphenate the element (but not it's children).
Implement two types of hyphenators:
-    language specific (hyphenate strings)
-    polyglot (hyphenate HTMLElements)

Notes:
With this, the PR can go in a minor release. Major release is to early.
Remove this functionality in next major release.

Makes createHyphenator private.

Fixes #108, [skip travis-ci]