feat: prevent small words and eol
Problem:
Some languages (e.g. polish) require not to have single letter words
at the end of a line.

Solution:
Implement a optional class based setting, find small words and replace
the space after them with a nbsp.
This is easy if we just search for single letter words, but get somehow
complicated when we open to any number of chars because we'll have
handle overlapping regex matches.

Notes:
First implementation and test
[skip travis-ci]