fix: Don't hyphenate words with COMBINING ACCENTs
Problem:
The current implementation for finding words first looks for consecutive
characters of the unicode property escape \p{Letter} and then checks if
this set contains a character that is not in the alphabet defined by the
.wasm file of the respective language.
Since combining accents are not part of \p{Letter} and there is no
normalized character the current implementation finds "БЕЛАР" and
hyphenates just this part of the word. This could lead to errors.

Solution:
Include subset of \p{Mn} in the word finding regex.
Currently the subset is: \u0300-\u036F\u0483-\u0487

Fixes #147