refactor: Subarray -> new TypedArray
Use new TypedArray instead of subarray for reading the hyphenated word
from memory.

In consequence return the length of the hyphenated word from wasm.hyphenate
(or 0 on error).

This saves some code bytes and is sligthly faster.

Since there is no documented use of standalone .wasm I don't consider this
as a breaking change.